gpg.py: Added a gpgme implementation of UIDExport (not yet functional).
We are calling UIDExport from GetKeySection.sign_key_async method and we do that for every UID of a key that we want to sign. Therefore , for each UID, the call to UIDExport does the following:
 * imports the key into a temp keyring
 * removes all other UIDs except the current one passed
 * returns the exported keydata with the UIDs removed
This is how we sign all UIDs and email back the key with its corresponding UID.
Unfortunately PyGPGME API doesn't seem to offer an easy way to delete an UID (or I don't know yet) so for know we don't use the UIDExport_gpgme implementation.