gpg: Some changes which give the library functions a more proper use.
This prepares the code for a "sign key" function which will have to do the next
steps:

1. Get the 'Primary Key' from the user's keyring, which will be used to sign.
2. For each UID of the key that must be signed:
    - check the sigs of that UID if it wasn't already signed by our key
    - if the UID matches the received UID , sign it using 'edit_sign' function
 from the gpgme.editutil module
    - return because we don't want to sign all UIDs

If we want to adhere to our current workflow, we need to call this function for
each UID of a key, and each call should import the key into a new context.
In this way we can sign each UID independently, and then send a mail with the
signed key.

We do it like this for now as we don't yet have a 'del_uid' function.