gpg: gpg_copy_secrets now copies the user's secring.
This is needed because we can't sign a key without a primary key inside our temporary
keyring. Unfortunately there is no '-export-secret-keys' option in gpgme API.

The other way we could go was to import the key into the user's own keyring and
delete it after we'd done with signing and exporting the key back.
But this would have raised another problem, if the user already has the key in his
keyring, then we would just delete it forever.