Changed gpg_copy_secrets to gpg_export_private_key.
This also changes the behavior of the function, now it will export the user's
private key into a temporary file which will be immediately deleted after we
import from the key into a temporary keyring. In this way we work in a temporar
directory so we don't import and sign keys inside user's default keyring.

While people don't like this way of exporting private key, we lack support from
GPGME which doesn't offer this option.