Eliminated a possible filesystem leak by exporting the private key to stdout
Changed the function's name to reflect more what it does, it imports the user's
private key into a temporary context. It can also receive a secret_key as
argument now and it will import from that instead of user's default keyring.

Removed extra argument 'gpg_homedir' as it only was used with logging.