gpg: gpg_get_keylist can receive extra argument for key expired
If set to True the function will return expired keys also,
else it won't. This is how we can implement the check for
expiration, because the gpgme module that we use doesn't recognize
the expired keys (aka the key.expired is always set to False).

Also, the 'expiration date' can only be retrieved from the
key.subkeys[0].expires field ... wierdly!