gpg: Add `export_secret_keys` function.
This separates the logic between importing and exporting private
keys.
The `export_secret_keys` function will return a string with all the
private keys found in the current gpg homedir. This is usefull for
when we want to sign a key, and we need to import our secret key
into a temporary keyring.